+++ lib/lacerda/contract.rb
@@ -18,6 +18,13 @@ module Lacerda
       end
     end
 
+    def object(name)
+      scoped_name = [service.name.underscore, name].join(Lacerda::SCOPE_SEPARATOR)
+      schema = @schema[:definitions][scoped_name.to_s]
+      raise Lacerda::Service::InvalidObjectTypeError unless schema
+      object_description_class.new(service, scoped_name, schema)
+    end
+
     private
 
     def load_schema(schema_or_file)