+++ lib/lacerda/service.rb
@@ -1,4 +1,5 @@
 require 'active_support/core_ext/string'
+require 'lacerda/service/error'
 
 module Lacerda
   # Models a service and its published objects as well as consumed
@@ -51,6 +52,18 @@ module Lacerda
       end
       print "\n" if verbose
       @errors.empty?
+    end
+
+    def validate_object_to_publish(type, data)
+      validate_object_to_publish!(type, data)
+      true
+    rescue
+      false
+    end
+
+    def validate_object_to_publish!(type, data)
+      object_description = @publish.object(type)
+      object_description.validate_data!(data)
     end
 
     private