+++ lib/lacerda/service/error.rb
@@ -1,0 +1,5 @@
+module Lacerda
+  class Service
+    class InvalidObjectTypeError < StandardError; end
+  end
+end