+++ CHANGELOG.markdown
@@ -1,3 +1,9 @@
+# 0.3.5 (29-Oct-15)
+- validation helpers for published objects
+
+# 0.3.4 (29-Oct-15)
+- enhance some error messages
+
 # 0.3.3 (28-Oct-15)
 - add error messages